Move 'pagination' to the core, barrel files for core
