+++ demo-shell-ng2/app/components/files/files.component.html
@@ -92,10 +92,10 @@
                 </content-action>
             </content-actions>
         </alfresco-document-list>
-        <alfresco-document-list-pagination
+        <alfresco-pagination
             [provider]="documentList.data"
             [supportedPageSizes]="[5, 10, 15, 20]">
-        </alfresco-document-list-pagination>
+        </alfresco-pagination>
     </alfresco-upload-drag-area>
 </div>
