+++ ng2-components/ng2-alfresco-core/index.ts
@@ -15,37 +15,20 @@
  * limitations under the License.
  */
 
-import { AlfrescoSettingsService } from './src/services/AlfrescoSettingsService.service';
-import { AlfrescoTranslationLoader } from './src/services/AlfrescoTranslationLoader.service';
-import { AlfrescoTranslationService } from './src/services/AlfrescoTranslationService.service';
-import { AlfrescoPipeTranslate } from './src/services/AlfrescoPipeTranslate.service';
-import { AlfrescoAuthenticationService } from './src/services/AlfrescoAuthenticationService.service';
-import { AlfrescoContentService } from './src/services/AlfrescoContentService.service';
-import { ContextMenuService } from './src/services/context-menu.service';
-import { ContextMenuHolderComponent } from './src/components/context-menu-holder.component';
-import { ContextMenuDirective } from './src/components/context-menu.directive';
-import { MDL } from './src/material/MaterialDesignLiteUpgradeElement';
-import { AlfrescoMdlButtonDirective } from './src/material/mdl-button.directive';
-import { AlfrescoMdlMenuDirective } from './src/material/mdl-menu.directive';
+import {
+    AlfrescoSettingsService,
+    AlfrescoTranslationLoader,
+    AlfrescoTranslationService,
+    AlfrescoPipeTranslate,
+    AlfrescoAuthenticationService,
+    AlfrescoContentService
+} from './src/services/index';
 
-export * from './src/services/AlfrescoSettingsService.service';
-export * from './src/services/AlfrescoTranslationLoader.service';
-export * from './src/services/AlfrescoTranslationService.service';
-export * from './src/services/AlfrescoPipeTranslate.service';
-export * from './src/services/AlfrescoAuthenticationService.service';
-export * from './src/services/AlfrescoContentService.service';
+import { ContextMenuService } from './src/components/context-menu/context-menu.service';
 
-// Material Design Lite integration
-export * from './src/material/MaterialDesignLiteUpgradeElement';
-export * from './src/material/mdl-button.directive';
-export * from './src/material/mdl-menu.directive';
-
-export * from './src/services/context-menu.service';
-export * from './src/components/context-menu-holder.component';
-export * from './src/components/context-menu.directive';
-
-export * from './src/data/pagination-provider';
-export * from './src/utils/object-utils';
+export * from './src/services/index';
+export * from './src/components/index';
+export * from './src/utils/index';
 
 export const ALFRESCO_CORE_PROVIDERS: [any] = [
     AlfrescoAuthenticationService,
@@ -57,17 +40,3 @@ export const ALFRESCO_CORE_PROVIDERS: [any] = [
     ContextMenuService
 ];
 
-export const CONTEXT_MENU_PROVIDERS: [any] = [
-    ContextMenuService
-];
-
-export const CONTEXT_MENU_DIRECTIVES: [any] = [
-    ContextMenuHolderComponent,
-    ContextMenuDirective
-];
-
-export const MATERIAL_DESIGN_DIRECTIVES: [any] = [
-    MDL,
-    AlfrescoMdlButtonDirective,
-    AlfrescoMdlMenuDirective
-];
