+++ ng2-components/ng2-alfresco-core/package.json
@@ -6,13 +6,14 @@
   "scripts": {
     "clean": "rimraf dist node_modules typings",
     "typings": "typings install",
-    "build": "npm run tslint && typings install && rimraf dist && tsc && license-check",
+    "build": "npm run tslint && typings install && rimraf dist && tsc && npm run copy-static && license-check",
     "build:w": "npm run tslint && typings install && rimraf dist && npm run watch-task",
     "watch-task": "concurrently \"npm run tsc:w\" \"license-check\"",
     "tslint": "npm run tslint-src && npm run tslint-root",
     "tslint-src": "tslint -c tslint.json src/{,**/}**.ts",
     "tslint-root": "tslint -c tslint.json *.ts",
     "licensecheck": "license-check",
+    "copy-static": "cpx \"./src/**/*.{html,css}\" ./dist/src",
     "tsc": "tsc",
     "tsc:w": "tsc -w",
     "pretest": "npm run build",
