+++ ng2-components/ng2-alfresco-core/src/components/context-menu/context-menu-holder.component.ts
@@ -16,7 +16,7 @@
  */
 
 import { Component, HostListener } from '@angular/core';
-import { ContextMenuService } from './../services/context-menu.service';
+import { ContextMenuService } from './context-menu.service';
 
 @Component({
     selector:'context-menu-holder',
