+++ ng2-components/ng2-alfresco-core/src/components/context-menu/context-menu.directive.ts
@@ -16,7 +16,7 @@
  */
 
 import { Directive, Input, HostListener } from '@angular/core';
-import { ContextMenuService } from './../services/context-menu.service';
+import { ContextMenuService } from './context-menu.service';
 
 @Directive({
     selector:'[context-menu]'
