+++ ng2-components/ng2-alfresco-core/src/components/context-menu/context-menu.service.spec.ts
