+++ ng2-components/ng2-alfresco-core/src/components/material/MaterialDesignLiteUpgradeElement.ts
