+++ ng2-components/ng2-alfresco-core/src/components/material/index.ts
@@ -1,0 +1,30 @@
+/*!
+ * @license
+ * Copyright 2016 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { MDL } from './MaterialDesignLiteUpgradeElement';
+import { AlfrescoMdlButtonDirective } from './mdl-button.directive';
+import { AlfrescoMdlMenuDirective } from './mdl-menu.directive';
+
+export * from './MaterialDesignLiteUpgradeElement';
+export * from './mdl-button.directive';
+export * from './mdl-menu.directive';
+
+export const MATERIAL_DESIGN_DIRECTIVES: [any] = [
+    MDL,
+    AlfrescoMdlButtonDirective,
+    AlfrescoMdlMenuDirective
+];
