+++ ng2-components/ng2-alfresco-core/src/components/material/mdl-menu.directive.ts
