+++ ng2-components/ng2-alfresco-core/src/components/pagination/pagination-provider.ts
