+++ ng2-components/ng2-alfresco-core/src/services/index.ts
@@ -1,0 +1,23 @@
+/*!
+ * @license
+ * Copyright 2016 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export * from './AlfrescoSettingsService.service';
+export * from './AlfrescoTranslationLoader.service';
+export * from './AlfrescoTranslationService.service';
+export * from './AlfrescoPipeTranslate.service';
+export * from './AlfrescoAuthenticationService.service';
+export * from './AlfrescoContentService.service';
